# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Publication API",
        "description": "API pour gérer les publications (texte, images et vidéos) les commentaires et les likes",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Server local"
        },
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/likes": {
            "post": {
                "tags": [
                    "Likes"
                ],
                "summary": "Ajouter un like",
                "operationId": "3709a56cfcde42dbc13291b1688744f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "publication_id",
                                    "rate"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "publication_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rate": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Like ajouté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Like"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation"
                    }
                }
            }
        },
        "/likes/{likeId}": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Récupérer un like par son ID",
                "operationId": "83ba8fcb930f010ef8435225b1f67148",
                "parameters": [
                    {
                        "name": "likeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du like",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Like"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Like non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Likes"
                ],
                "summary": "Mettre à jour un like",
                "operationId": "a2ba2b5393cd13aaa799b3e571b62742",
                "parameters": [
                    {
                        "name": "likeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rate"
                                ],
                                "properties": {
                                    "rate": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Like mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Like"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur lors de la mise à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Likes"
                ],
                "summary": "Supprimer un like",
                "operationId": "8ee81ce17329c955099902c00ff5f1cb",
                "parameters": [
                    {
                        "name": "likeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Like supprimé avec succès"
                    },
                    "400": {
                        "description": "Erreur lors de la suppression"
                    }
                }
            }
        },
        "/publications/{publicationId}/likes": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Récupérer les likes d'une publication",
                "operationId": "9ceb2a6fc3ae6c2ff6a2b2de1ff4d58f",
                "parameters": [
                    {
                        "name": "publicationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des likes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Like"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Publication non trouvée"
                    }
                }
            }
        },
        "/users/{userId}/likes": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Récupérer les likes d'un utilisateur",
                "operationId": "7d5211bee57f421eced93ee8ca81123f",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des likes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Like"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Utilisateur non trouvé"
                    }
                }
            }
        },
        "/publications/{publicationId}/average-rate": {
            "get": {
                "tags": [
                    "Likes"
                ],
                "summary": "Récupérer la moyenne des notes d'une publication",
                "operationId": "5d63a3e4bddd57ccfbf06637ebbc3722",
                "parameters": [
                    {
                        "name": "publicationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Moyenne des notes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "average_rate": {
                                            "type": "number",
                                            "example": 4.5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun like trouvé pour cette publication"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Like": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "publication_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "rate": {
                        "type": "integer",
                        "example": 4
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}